Move tests and get core passing.
