+++ packages/rut-dom/tests/features/__mocks__/Lazy.js
