+++ packages/rut-dom/tests/features/errorBoundary.test.tsx
