+++ packages/rut-dom/tests/features/hoc.test.tsx
