+++ packages/rut-dom/tests/features/lazy.test.tsx
