+++ packages/rut/tests/helpers.ts
@@ -1,4 +1,17 @@
+import { createAdapter, Element, SyntheticEvent } from '../src/adapters';
 import { MatchResult } from '../src/types';
+
+class TestElement extends Element {
+  createSyntheticEvent(type: string) {
+    // @ts-ignore
+    return new SyntheticEvent(type, {
+      preventDefault() {},
+      stopPropagation() {},
+    });
+  }
+}
+
+export const { render, renderAndWait } = createAdapter(instance => new TestElement(instance));
 
 export function formatMatcherMessage(result: MatchResult, isNot: boolean) {
   return (isNot ? result.notMessage : result.message)
