+++ packages/rut/tests/matchers/toBeChecked.test.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
-import { render } from '../../src/render';
 import toBeChecked from '../../src/matchers/toBeChecked';
-import { runMatcher } from '../helpers';
+import { render, runMatcher } from '../helpers';
 
 describe('toBeChecked()', () => {
   it('errors if a non-Element is passed', () => {
