+++ packages/rut/tests/matchers/toBeDisabled.test.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
-import { render } from '../../src/render';
 import toBeDisabled from '../../src/matchers/toBeDisabled';
-import { runMatcher } from '../helpers';
+import { render, runMatcher } from '../helpers';
 
 describe('toBeDisabled()', () => {
   it('errors if a non-Element is passed', () => {
