+++ packages/rut/tests/matchers/toHaveClassName.test.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
-import { render } from '../../src/render';
 import toHaveClassName from '../../src/matchers/toHaveClassName';
-import { runMatcher } from '../helpers';
+import { render, runMatcher } from '../helpers';
 
 describe('toHaveClassName()', () => {
   it('errors if a non-Element is passed', () => {
