+++ packages/rut/tests/matchers/toHaveValue.test.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
-import { render } from '../../src/render';
 import toHaveValue from '../../src/matchers/toHaveValue';
-import { runMatcher } from '../helpers';
+import { render, runMatcher } from '../helpers';
 
 describe('toHaveValue()', () => {
   it('errors if a non-Element is passed', () => {
